--- frontend-react/e2e/pages/developer-resources/index.ts
@@ -1,17 +1,0 @@
-import { BasePage, BasePageTestArgs } from "../BasePage";
-
-export class DeveloperResourcesIndexPage extends BasePage {
-    constructor(testArgs: BasePageTestArgs) {
-        super(
-            {
-                url: "/developer-resources",
-                title: "ReportStream developer resources",
-                heading: testArgs.page.getByRole("heading", {
-                    name: "Developer resources",
-                    exact: true,
-                }),
-            },
-            testArgs,
-        );
-    }
-}
