+++ frontend-react/e2e/pages/developer-resources/resources.ts
